Finished implementing Scoreboard
All of the functions for Scoreboard have been implemented; it is possible to add new scores, clear the score, serialize, and export it. Some bug fixes in Double Linked List were also implemented, as well as changing the format of the print function in Score

However, there is an issue in deserializing the scoreboard_data. The deserialized date_time do not seem to show when printed.